feat: delete leadership key working, rename use-case
