+++ api/src/domains/teams/routes.js
@@ -12,7 +12,7 @@ const teamRoutes = () => {
 
     router.post('/', controller.addTeam)
     router.post('/:id/generate-lead-keys', controller.generateLeadKeys)
-    router.post('/delete-lead-key', controller.deleteLeadKey)
+    router.post('/delete-lead-key', controller.deleteLeadershipKey)
     router.get('/', controller.getTeams)
     router.get('/:id', controller.getTeamById)
     router.get('/:id/shifts', controller.getShifts)
